Season creation updates
Added ability to include creation of 'x' teams with 5 shooters on each
team when the season is created.
Added missed_weeks relation on seasons for holidays and postponment due
to say weather.
Added UI on Matches to add holidays before schedule creation.  This
also includes jquery-ui resources because I needed a datepicker.